Fix initial build directory used in build step
In the BuildStep constructor the build configuration is not
fully set up, leading to an empty build directory in there.

It is not needed that we update the command directly on changes
though (it is not visible in the projects mode UI anywhere),
so just initialize the build step in its init() method.

Change-Id: Iab168d15f1224c2e1ca66017244a80ed1059f417
Reviewed-by: Christian Stenger <christian.stenger@qt.io>